Update
This project is over for now. Now it is an working monopoly manager which is responsive on laptop, tab and mobile with all the features and transactions which the banker does can be done.
Along with that there is a provision for seeing each player's statistics at the current situation of the game like an analysis . Lastly it can store your last played game for you and can also give the winner and the standings of the players in one go whenever you want.
enjoy:)